F4.cpp followup -- other files are affected by mingling int32_t, a typedef, with other types through overloading
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@4469 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/RR.cpp
@@ -98,7 +98,7 @@ void RingRR::remove_elem(RRelem f) const
 ring_elem RingRR::random() const
 {
   int d = 1000000;
-  long r = Random::random0(2*d);
+  long r = Random::random0((int32_t)(2*d));
   double a = (r*1.0)/d - 1.0;
   return RingRR::from_double(a);
 }
